+++ pkg/util/validation/limits.go
@@ -4,10 +4,12 @@ import (
 	"encoding/json"
 	"errors"
 	"flag"
+	"math"
 	"time"
 
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/relabel"
+	"golang.org/x/time/rate"
 
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 )
@@ -99,6 +101,10 @@ type Limits struct {
 	// Alertmanager.
 	AlertmanagerReceiversBlockCIDRNetworks     flagext.CIDRSliceCSV `yaml:"alertmanager_receivers_firewall_block_cidr_networks" json:"alertmanager_receivers_firewall_block_cidr_networks"`
 	AlertmanagerReceiversBlockPrivateAddresses bool                 `yaml:"alertmanager_receivers_firewall_block_private_addresses" json:"alertmanager_receivers_firewall_block_private_addresses"`
+
+	// Alertmanager limits
+	EmailNotificationRateLimit float64 `yaml:"alertmanager_email_notification_rate_limit" json:"alertmanager_email_notification_rate_limit"`
+	EmailNotificationBurstSize int     `yaml:"alertmanager_email_notification_burst_size" json:"alertmanager_email_notification_burst_size"`
 }
 
 // RegisterFlags adds the flags required to config this to the given FlagSet
@@ -159,6 +165,8 @@ func (l *Limits) RegisterFlags(f *flag.FlagSet) {
 	// Alertmanager.
 	f.Var(&l.AlertmanagerReceiversBlockCIDRNetworks, "alertmanager.receivers-firewall-block-cidr-networks", "Comma-separated list of network CIDRs to block in Alertmanager receiver integrations.")
 	f.BoolVar(&l.AlertmanagerReceiversBlockPrivateAddresses, "alertmanager.receivers-firewall-block-private-addresses", false, "True to block private and local addresses in Alertmanager receiver integrations. It blocks private addresses defined by  RFC 1918 (IPv4 addresses) and RFC 4193 (IPv6 addresses), as well as loopback, local unicast and local multicast addresses.")
+	f.Float64Var(&l.EmailNotificationRateLimit, "alertmanager.email-notification-rate-limit", 0, "Per-user rate limit for sending email notifications from Alertmanager in emails/sec. 0 = rate limit disabled. Negative value = no emails are allowed.")
+	f.IntVar(&l.EmailNotificationBurstSize, "alertmanager.email-notification-burst-size", 1, "Per-user burst size for email notifications. If set to 0, no email notifications will be sent, unless rate-limit is disabled, in which case all email notifications are allowed.")
 }
 
 // Validate the limits config and returns an error if the validation
@@ -498,6 +506,26 @@ func (o *Overrides) AlertmanagerReceiversBlockCIDRNetworks(user string) []flagex
 // in the Alertmanager receivers for the given user.
 func (o *Overrides) AlertmanagerReceiversBlockPrivateAddresses(user string) bool {
 	return o.getOverridesForUser(user).AlertmanagerReceiversBlockPrivateAddresses
+}
+
+func (o *Overrides) EmailNotificationRateLimit(user string) rate.Limit {
+	l := o.getOverridesForUser(user).EmailNotificationRateLimit
+	if l == 0 || math.IsInf(l, 1) {
+		return rate.Inf // No rate limit.
+	}
+
+	if l < 0 {
+		l = 0 // No emails will be sent.
+	}
+	return rate.Limit(l)
+}
+
+func (o *Overrides) EmailNotificationBurst(user string) int {
+	b := o.getOverridesForUser(user).EmailNotificationBurstSize
+	if b < 0 {
+		b = 0
+	}
+	return b
 }
 
 func (o *Overrides) getOverridesForUser(userID string) *Limits {